+++ modules/canvaskit/CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    The fastest choice for gradients is the flat Float32Array, the fastest choice for drawAtlas and
    MakeSkVertices is the flat Uint32Array.
  - Color arrays may also be objects created with CanvasKit.Malloc
+ - renamed `reportBackendType` to `reportBackendTypeIsGPU` and made it return a boolean
 
 ### Fixed
  - `TextStyle.color` can correctly be a Malloc'd Float32Array.